requirements.txt
qrcode
io
Crypto.Cipher
base64
random
flask
gunicorn